+++ Fit&Nut/PlaniProgramModul/src/main/java/com/planiprogram/repository/OsobaRepository.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import com.planiprogram.module.Osoba;
 
 @RepositoryRestResource(path="osobe",collectionResourceRel="osobe")
-public interface OsobaRepository extends PagingAndSortingRepository<Osoba, Long>{
+public interface OsobaRepository extends PagingAndSortingRepository<Osoba, Integer>{
 
 }
